Couple of typos on the README
I read the entire readme and spotted a few typos. Maybe it is a bit early for this kind of work, but well, they don't hurt :P

I am not really into node yet to understand too much the implementation, but seems really easy to work with.

Could it work with koa? Or maybe it needs some wrapping (from total ignorance).